requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pymongo>=4.0
httpx>=0.27.0
pdfminer.six>=20221105
python-docx>=1.0.0
